requirements.txt
pymupdf
sentence-transformers
faiss-cpu
ollama
numpy
streamlit
chromadb
google-generativeai
python-dotenv